refactor: rename openai/utils/tokens.py to utils/openai_token.py
