+++ pyproject.toml
@@ -32,6 +32,7 @@ dev-dependencies = [
     "ruff>=0.1.3",
     "pytest>=7.4.3",
     "langchain>=0.0.27",
+    "openai>=1.3.2",
 ]
 
 [tool.rye.scripts]
