+++ requirements-dev.lock
@@ -18,9 +18,13 @@ certifi==2023.7.22
 charset-normalizer==3.3.1
 dataclasses-json==0.6.1
 deprecated==1.2.14
+distro==1.8.0
 exceptiongroup==1.1.3
 frozenlist==1.4.0
 googleapis-common-protos==1.61.0
+h11==0.14.0
+httpcore==1.0.2
+httpx==0.25.1
 idna==3.4
 importlib-metadata==6.8.0
 iniconfig==2.0.0
@@ -32,6 +36,7 @@ marshmallow==3.20.1
 multidict==6.0.4
 mypy-extensions==1.0.0
 numpy==1.24.4
+openai==1.3.2
 opentelemetry-api==1.20.0
 opentelemetry-exporter-otlp-proto-common==1.20.0
 opentelemetry-exporter-otlp-proto-http==1.20.0
@@ -53,6 +58,7 @@ sqlalchemy==2.0.22
 tenacity==8.2.3
 tiktoken==0.5.1
 tomli==2.0.1
+tqdm==4.66.1
 typing-extensions==4.8.0
 typing-inspect==0.9.0
 urllib3==2.0.7
