+++ tests/utils/test_openai_token.py
@@ -1,4 +1,4 @@
-from greptimeai.openai.utils.tokens import (
+from greptimeai.utils.openai_token import (
     cal_openai_token_cost_for_model,
     num_tokens_from_messages,
 )
